third_party/blink/web_tests/platform/generic/fast/events/init-message-event-expected.txt

Checks that initMessageEvent() is working

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS newEvent.initMessageEvent("message", true, true, "ff", "*", 43, window, {x:1}); threw exception TypeError: Failed to execute 'initMessageEvent' on 'MessageEvent': The object must have a callable @@iterator property..
PASS newEvent.initMessageEvent("message", true, true, "ff", "*", 43, window, undefined) did not throw exception.
PASS [object MessageEvent] is an instance of function MessageEvent() { [native code] }
PASS successfullyParsed is true

TEST COMPLETE
